yn-organicperf-suite/requirements.txt
# Core
streamlit>=1.30.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
aiohttp>=3.9.0

# NLP / ML (Semantic Score module)
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0
nltk>=3.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-Levenshtein>=0.23.0

# Web scraping
beautifulsoup4>=4.12.0
trafilatura>=1.6.0
readability-lxml>=0.8.0
lxml>=4.9.0

# Language detection
langdetect>=1.0.9

# OpenAI
openai>=1.10.0
httpx>=0.25.0

# Export
xlsxwriter>=3.1.0
